Final day, first commit

+++ app/src/main/java/kr/ac/inha/nsl/mindnavigator/InterventionsActivity.java
@@ -29,7 +29,7 @@ public class InterventionsActivity extends AppCompatActivity {
     //region Variables
     boolean saveIntervention = false;
     static String result = null;
-    static int resultSchedule = 0;
+    static short resultSchedule = 0;
 
     EditText interv_text;
     View interv_choice;
@@ -264,7 +264,7 @@ public class InterventionsActivity extends AppCompatActivity {
                                         String interv_name = (String) args[0];
                                         Toast.makeText(InterventionsActivity.this, "Intervention successfully created!", Toast.LENGTH_SHORT).show();
                                         result = interv_name;
-                                        resultSchedule = Integer.parseInt((String) intervScheduling.findViewById(intervScheduling.getCheckedRadioButtonId()).getTag());
+                                        resultSchedule = Short.parseShort((String) intervScheduling.findViewById(intervScheduling.getCheckedRadioButtonId()).getTag());
                                         setResult(Activity.RESULT_OK);
                                         finish();
                                         overridePendingTransition(R.anim.activity_in_reverse, R.anim.activity_out_reverse);
@@ -280,7 +280,7 @@ public class InterventionsActivity extends AppCompatActivity {
                                         String interv_name = (String) args[0];
                                         Toast.makeText(InterventionsActivity.this, "Intervention already exists. Thus, it was picked for you.", Toast.LENGTH_SHORT).show();
                                         result = interv_name;
-                                        resultSchedule = Integer.parseInt((String) intervScheduling.findViewById(intervScheduling.getCheckedRadioButtonId()).getTag());
+                                        resultSchedule = Short.parseShort((String) intervScheduling.findViewById(intervScheduling.getCheckedRadioButtonId()).getTag());
                                         setResult(Activity.RESULT_OK);
                                         finish();
                                         overridePendingTransition(R.anim.activity_in_reverse, R.anim.activity_out_reverse);
@@ -309,7 +309,7 @@ public class InterventionsActivity extends AppCompatActivity {
                 Toast.makeText(this, "Please pick an intervention first!", Toast.LENGTH_SHORT).show();
                 return;
             }
-            resultSchedule = Integer.parseInt((String) intervScheduling.findViewById(intervScheduling.getCheckedRadioButtonId()).getTag());
+            resultSchedule = Short.parseShort((String) intervScheduling.findViewById(intervScheduling.getCheckedRadioButtonId()).getTag());
             setResult(Activity.RESULT_OK);
             finish();
             overridePendingTransition(R.anim.activity_in_reverse, R.anim.activity_out_reverse);
